refactor: move stores to their own files
