+++ parachat/src/byondcalls/admin.ts
@@ -1,5 +1,5 @@
-import { useAdminSlice } from '~/common/store';
 import { ByondCall } from '~/common/types';
+import { useAdminSlice } from '~/stores/admin';
 
 export const playAudio: ByondCall = topic => {
   useAdminSlice.getState().setCurrentAudio(topic);
