+++ parachat/src/components/settings/SettingsMenu.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { Transition } from 'react-transition-group';
 import styled from 'styled-components';
 import { animationDurationMs, fadeIn, slideUp } from '~/common/animations';
-import { useHeaderSlice } from '~/common/store';
 import { SettingsTab } from '~/common/types';
+import { useHeaderSlice } from '~/stores/header';
 import Categories from './Categories';
 import GeneralSettings from './categories/General';
 import HighlightSettings from './categories/Highlight';
